tests/ce2014a_vkeys.html

<!doctype html>
<html lang="en">
<head>
<meta charset="utf-8">

<title>Generated Key Events</title>
<!--[if lt IE 9]>
<script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
<![endif]-->
</head>
	
<script src="../resources/testharness.js"></script>

<body>
<script>

test(function() {
	assert_equals(KeyEvent.VK_CANCEL, 3);
}, "VK_CANCEL is equal to 3.");

</script>
<div id="keyTarget"></div>

</body>
</html>
